refactor: [#141] 반환 타입 수정

+++ green/src/main/java/com/greengrim/green/core/challenge/service/GetChallengeService.java
@@ -156,7 +156,7 @@ public class GetChallengeService {
             Optional<ChatMessage> chatMessage = chatRepository.
                 findFirstByRoomIdOrderByCreatedAtDesc(chatroomId);
 
-            int newMessageCount = chatRepository.
+            int newMessageCount = (int) chatRepository.
                 countByRoomIdAndCreatedAtAfter(challenge.getChatroom().getId(),
                     visitMap.get(chatroomId));
